+++ commitlog/recovery.go
@@ -267,6 +267,9 @@ func nextSequenceValueForTable(table *store.Table) (uint64, bool) {
 }
 
 func autoIncrementValueAsUint64(v types.Value, kind schema.ValueKind) (uint64, bool) {
+	if v.IsNull() {
+		return 0, false
+	}
 	switch kind {
 	case schema.KindInt8:
 		n := int64(v.AsInt8())